The xBlinds project started because I wanted a (cheap) way to motorize my vertical blinds
The xBlinds project started because I wanted a (cheap) way to motorize my vertical blinds and existing commercial solutions are heavily priced and not getting the best reviews, so a wintery night I started sketching the UI in Notepad.

This project is two-fold; the bin and UI published here and the STL for the casing going up on Thingiverse.